Refactor: Split the Jenkins Typer commands
typer_apps/jenkins.py held six unrelated command groups in 754 lines:
server-wide commands, builds, jobs, nodes, plugins and tokens. The six
Typer apps were declared together at the top and wired together there
too, while the commands they carried were scattered down the file, so
finding the commands belonging to one group meant scrolling past the
other five.

Turn the module into a package where each group owns its own module
and declares its own Typer app. The package __init__ composes them,
which is now the only place the command tree is described.

Each group keeps its own helpers alongside the commands that use them:
the job enable/disable Groovy template, the node status formatter, the
plugin table helpers, and the jenkins_jobs.ini guard with the
credentials message. None was shared across groups.

The composition uses ordinary imports rather than imports kept solely
for their registration side effect, so nothing here needs a linter
exemption.

The command tree is unchanged: all 197 commands and their options
compare identically before and after.
